move tui `Mode` and related types into their own module
