+++ crates/but/src/command/legacy/status/tui/mod.rs
@@ -6,7 +6,6 @@ use but_ctx::Context;
 use but_rebase::graph_rebase::mutate::InsertSide;
 use crossterm::event::{self, Event, KeyCode, KeyEvent};
 use gitbutler_operating_modes::OperatingMode;
-use gitbutler_stack::StackId;
 use itertools::Either;
 use ratatui::{
     Frame,
@@ -31,11 +30,14 @@ use crate::{
                 graph_extension::{ExtensionDirection, extend_connector_spans},
                 highlight::{Highlights, with_highlight},
                 key_bind::{KeyBinds, confirm_key_binds, default_key_binds},
+                mode::{
+                    CommandMode, CommitMode, CommitSource, InlineRewordMode, Mode, MoveMode,
+                    MoveSource, RubMode,
+                },
                 toast::{ToastKind, Toasts},
             },
         },
     },
-    id::{ShortId, UncommittedCliId},
     tui::{CrosstermTerminalGuard, TerminalGuard},
     utils::{DebugAsType, OutputChannel},
 };
@@ -51,6 +53,7 @@ mod details;
 mod graph_extension;
 mod highlight;
 mod key_bind;
+mod mode;
 mod operations;
 mod rub_api;
 mod toast;
@@ -2359,224 +2362,6 @@ enum BranchMessage {
 enum FilesMessage {
     ToggleGlobalFilesList,
     ToggleFilesForCommit,
-}
-
-#[derive(Debug, Default, strum::EnumDiscriminants)]
-#[strum_discriminants(derive(strum::EnumIter, Hash))]
-#[strum_discriminants(name(ModeDiscriminant))]
-enum Mode {
-    #[default]
-    Normal,
-    Rub(RubMode),
-    RubButApi(RubMode),
-    InlineReword(InlineRewordMode),
-    Command(CommandMode),
-    Commit(CommitMode),
-    Move(MoveMode),
-    Branch,
-}
-
-#[derive(Debug)]
-struct RubMode {
-    source: Arc<CliId>,
-    available_targets: Vec<Arc<CliId>>,
-}
-
-#[derive(Debug)]
-enum InlineRewordMode {
-    Commit {
-        commit_id: gix::ObjectId,
-        textarea: Box<TextArea<'static>>,
-    },
-    Branch {
-        name: String,
-        stack_id: StackId,
-        textarea: Box<TextArea<'static>>,
-    },
-}
-
-impl InlineRewordMode {
-    fn textarea(&self) -> &TextArea<'static> {
-        match self {
-            InlineRewordMode::Commit { textarea, .. }
-            | InlineRewordMode::Branch { textarea, .. } => textarea,
-        }
-    }
-
-    fn textarea_mut(&mut self) -> &mut TextArea<'static> {
-        match self {
-            InlineRewordMode::Commit { textarea, .. }
-            | InlineRewordMode::Branch { textarea, .. } => textarea,
-        }
-    }
-}
-
-#[derive(Debug)]
-struct CommandMode {
-    textarea: Box<TextArea<'static>>,
-}
-
-#[derive(Debug)]
-struct CommitMode {
-    source: Arc<CommitSource>,
-    /// If set, then the commit must be made on this stack
-    ///
-    /// Used when committing changes staged to a specific stack
-    scope_to_stack: Option<StackId>,
-    /// The side to insert the new commit on, relative to the target commit.
-    ///
-    /// Note this is only respected when inserting at a commit. If inserting at a branch we'll
-    /// always use [`InsertSide::Below`].
-    insert_side: InsertSide,
-}
-
-/// A subset of [`CliId`] that supports being committed
-#[derive(Debug)]
-enum CommitSource {
-    Unassigned(UnassignedCommitSource),
-    Uncommitted(Box<UncommittedCliId>),
-    Stack(StackCommitSource),
-}
-
-#[derive(Debug)]
-struct UnassignedCommitSource {
-    id: ShortId,
-}
-
-#[derive(Debug)]
-struct StackCommitSource {
-    id: ShortId,
-    stack_id: StackId,
-}
-
-impl TryFrom<CliId> for CommitSource {
-    type Error = anyhow::Error;
-
-    fn try_from(id: CliId) -> Result<Self, Self::Error> {
-        match id {
-            CliId::Unassigned { id } => Ok(Self::Unassigned(UnassignedCommitSource { id })),
-            CliId::Uncommitted(uncommitted_cli_id) => {
-                Ok(Self::Uncommitted(Box::new(uncommitted_cli_id)))
-            }
-            CliId::Stack { id, stack_id } => Ok(Self::Stack(StackCommitSource { id, stack_id })),
-            CliId::PathPrefix { .. }
-            | CliId::CommittedFile { .. }
-            | CliId::Branch { .. }
-            | CliId::Commit { .. } => anyhow::bail!("cannot commit: {id:?}"),
-        }
-    }
-}
-
-impl PartialEq<CliId> for CommitSource {
-    fn eq(&self, other: &CliId) -> bool {
-        match self {
-            CommitSource::Unassigned(UnassignedCommitSource { id: lhs_id }) => {
-                if let CliId::Unassigned { id: rhs_id } = other {
-                    lhs_id == rhs_id
-                } else {
-                    false
-                }
-            }
-            CommitSource::Uncommitted(lhs) => {
-                if let CliId::Uncommitted(rhs) = other {
-                    &**lhs == rhs
-                } else {
-                    false
-                }
-            }
-            CommitSource::Stack(StackCommitSource {
-                id: id_lhs,
-                stack_id: stack_id_lhs,
-            }) => {
-                if let CliId::Stack {
-                    id: id_rhs,
-                    stack_id: stack_id_rhs,
-                } = other
-                {
-                    id_lhs == id_rhs && stack_id_lhs == stack_id_rhs
-                } else {
-                    false
-                }
-            }
-        }
-    }
-}
-
-#[derive(Debug)]
-struct MoveMode {
-    source: Arc<MoveSource>,
-    insert_side: InsertSide,
-}
-
-/// A subset of [`CliId`] that supports being moved
-#[derive(Debug)]
-enum MoveSource {
-    Commit {
-        commit_id: gix::ObjectId,
-        id: ShortId,
-    },
-    Branch {
-        name: String,
-        id: ShortId,
-        stack_id: Option<StackId>,
-    },
-}
-
-impl TryFrom<CliId> for MoveSource {
-    type Error = anyhow::Error;
-
-    fn try_from(id: CliId) -> Result<Self, Self::Error> {
-        match id {
-            CliId::Branch { name, id, stack_id } => Ok(Self::Branch { name, id, stack_id }),
-            CliId::Commit { commit_id, id } => Ok(Self::Commit { commit_id, id }),
-            CliId::Uncommitted(uncommitted_cli_id) => {
-                anyhow::bail!("cannot move: {:?}", uncommitted_cli_id.id)
-            }
-            CliId::PathPrefix { id, .. }
-            | CliId::CommittedFile { id, .. }
-            | CliId::Unassigned { id }
-            | CliId::Stack { id, .. } => {
-                anyhow::bail!("cannot move: {id:?}")
-            }
-        }
-    }
-}
-
-impl PartialEq<CliId> for MoveSource {
-    fn eq(&self, other: &CliId) -> bool {
-        match self {
-            MoveSource::Commit {
-                commit_id: commit_id_lhs,
-                id: id_lhs,
-            } => {
-                if let CliId::Commit {
-                    commit_id: commit_id_rhs,
-                    id: id_rhs,
-                } = other
-                {
-                    commit_id_lhs == commit_id_rhs && id_lhs == id_rhs
-                } else {
-                    false
-                }
-            }
-            MoveSource::Branch {
-                name: name_lhs,
-                id: id_lhs,
-                stack_id: stack_id_lhs,
-            } => {
-                if let CliId::Branch {
-                    name: name_rhs,
-                    id: id_rhs,
-                    stack_id: stack_id_rhs,
-                } = other
-                {
-                    name_lhs == name_rhs && id_lhs == id_rhs && stack_id_lhs == stack_id_rhs
-                } else {
-                    false
-                }
-            }
-        }
-    }
 }
 
 /// What to select after reloading
